refactor : BaseTime 도메인 공통 패키지로 변경
