+++ src/main/java/com/jasik/momsnaggingapi/domain/diary/Diary.java
@@ -1,6 +1,6 @@
 package com.jasik.momsnaggingapi.domain.diary;
 
-import com.jasik.momsnaggingapi.domain.BaseTime;
+import com.jasik.momsnaggingapi.domain.common.BaseTime;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
